made shift right and left in editor work slightly better

+++ src/drawing.c
@@ -284,9 +284,9 @@ void drawBars()
 {
 	//Draw the bars
 	float middle = GetScreenWidth()/2;
-	float distBetweenBeats = getMusicDuration() / getBeatsCount() / _barMeasureCount;
+	double distBetweenBeats = (double)getMusicDuration() / getBeatsCount() / _barMeasureCount;
 
-	float distBetweenBars = distBetweenBeats*4;
+	double distBetweenBars = distBetweenBeats*4;
 	for (int i = (screenToMusicTime(0)-_map->offset/1000.0)/distBetweenBars; i < (screenToMusicTime(GetScreenWidth())-_map->offset/1000.0)/distBetweenBars; i++)
 	{
 		DrawRectangle(musicTimeToScreen(distBetweenBars*i+_map->offset/1000.0)-10,GetScreenHeight()*0.6,GetScreenWidth()*0.01,GetScreenHeight()*0.3,(Color){.r=255,.g=255,.b=255,.a=180});
